Fix chart components, reducing firebase requests, and show total spent by category on pie chart
